ecommerce/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
stripe==13.1.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
